move methods to tseViewApplication
