+++ build.gradle
@@ -300,7 +300,7 @@ dependencyManagement {
 
 dependencies {
     implementation group: 'com.github.hmcts', name: 'et-common', version:'1.0.15'
-    implementation group: 'com.github.hmcts', name: 'et-data-model', version:'1.8.0'
+    implementation group: 'com.github.hmcts', name: 'et-data-model', version:'1.8.2'
     implementation group: 'com.github.hmcts', name: 'ccd-case-document-am-client', version: '1.7.3'
     implementation group: 'com.github.hmcts', name: 'document-management-client', version: '7.0.1'
     implementation group: 'com.github.hmcts.java-logging', name: 'logging', version: reformLoggingVersion
