+++ src/test/java/uk/gov/hmcts/ethos/replacement/docmosis/helpers/TseHelperTest.java
@@ -29,7 +29,12 @@ import java.util.UUID;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
 import static org.junit.Assert.assertNull;
-import static uk.gov.hmcts.ecm.common.model.helper.Constants.*;
+import static uk.gov.hmcts.ecm.common.model.helper.Constants.CLAIMANT_TITLE;
+import static uk.gov.hmcts.ecm.common.model.helper.Constants.NO;
+import static uk.gov.hmcts.ecm.common.model.helper.Constants.OPEN_STATE;
+import static uk.gov.hmcts.ecm.common.model.helper.Constants.RESPONDENT_TITLE;
+import static uk.gov.hmcts.ecm.common.model.helper.Constants.TSE_APP_POSTPONE_A_HEARING;
+import static uk.gov.hmcts.ecm.common.model.helper.Constants.YES;
 
 @SuppressWarnings({"PMD.LinguisticNaming", "PMD.ExcessiveImports"})
 public class TseHelperTest {
@@ -77,28 +82,6 @@ public class TseHelperTest {
                     .build())
                 .build()));
         DynamicFixedListType actual = TseHelper.populateRespondentSelectApplication(caseData);
-        assertThat(actual.getListItems().size(), is(0));
-    }
-
-    @Test
-    public void populateOpenOrClosedApplications_withEmptyList_doesNothing() {
-        caseData.setGenericTseApplicationCollection(null);
-        caseData.setTseViewApplicationOpenOrClosed("Open");
-        DynamicFixedListType actual = TseHelper.populateOpenOrClosedApplications(caseData);
-        assertNull(actual);
-    }
-
-    @Test
-    public void populateOpenApplications_withAnOpenApplication_returnsDynamicList() {
-        caseData.setTseViewApplicationOpenOrClosed("Open");
-        DynamicFixedListType actual = TseHelper.populateOpenOrClosedApplications(caseData);
-        assertThat(actual.getListItems().size(), is(1));
-    }
-
-    @Test
-    public void populateClosedApplications_withNoClosedApplications_returnEmptyList() {
-        caseData.setTseViewApplicationOpenOrClosed("Closed");
-        DynamicFixedListType actual = TseHelper.populateOpenOrClosedApplications(caseData);
         assertThat(actual.getListItems().size(), is(0));
     }
 
@@ -298,197 +281,5 @@ public class TseHelperTest {
         );
 
         assertThat(actual.toString(), is(expected.toString()));
-    }
-
-    private DocumentTypeItem createDocumentTypeItem(String fileName) {
-        DocumentTypeItem documentTypeItem = new DocumentTypeItem();
-        documentTypeItem.setId("1234");
-        documentTypeItem.setValue(DocumentTypeBuilder.builder().withUploadedDocument(fileName, "1234").build());
-        return documentTypeItem;
-    }
-
-//    @Test
-//    public void createResponseTable_ReturnString() {
-//
-//        TseRespondType tseRespondType = TseRespondType.builder()
-//                .from(CLAIMANT_TITLE)
-//                .date("23 December 2022")
-//                .response("Response Details")
-//                .hasSupportingMaterial(YES)
-//                .supportingMaterial(List.of(
-//                        createDocumentTypeItem("image.png"),
-//                        createDocumentTypeItem("Form.pdf")))
-//                .copyToOtherParty(NO)
-//                .build();
-//
-//        TseRespondTypeItem tseRespondTypeItem = TseRespondTypeItem.builder()
-//                .id(UUID.randomUUID().toString())
-//                .value(tseRespondType)
-//                .build();
-//
-//        List<TseRespondTypeItem> respondList = List.of(tseRespondTypeItem);
-//        String applicant = "Respondent";
-//
-//        String fileDisplay1 = "<a href=\"/documents/1234/binary\" target=\"_blank\">image.png</a>";
-//
-//        String fileDisplay2 = "<a href=\"/documents/1234/binary\" target=\"_blank\">Form.pdf</a>";
-//
-//        String expected = "|Responses | |\r\n"
-//                + "|--|--|\r\n\r\n"
-//                + "|Response 1 | |\r\n"
-//                + "|--|--|\r\n"
-//                + "|Response from | Claimant|\r\n"
-//                + "|Response date | 23 December 2022|\r\n"
-//                + "|What’s your response to the respondent’s application? | Response Details|\r\n"
-//                + "|Supporting material | " + fileDisplay1 + "<br>" + fileDisplay2 + "<br>" + "|\r\n"
-//                + "|Do you want to copy this correspondence to the other party to satisfy the Rules of Procedure? "
-//                + "| No|\r\n"
-//                + "|Details of why you do not want to inform the other party | N/A|\r\n"
-//                + "\r\n";
-//
-//        TseHelper.setDataForTseApplicationSummaryAndResponses(caseData);
-//        assertThat(caseData.getTseApplicationSummaryAndResponsesMarkup(), is(expected));
-//    }
-
-    @Test
-    public void setDataForTseApplicationSummaryAndResponses_withEmptyList_doesNothing() {
-        caseData.setGenericTseApplicationCollection(null);
-        TseHelper.setDataForTseApplicationSummaryAndResponses(caseData);
-        assertNull(caseData.getTseApplicationSummaryAndResponsesMarkup());
-    }
-
-    @Test
-    public void setDataForTseApplicationSummaryAndResponses_withAnApplication_setsApplicationSummary() {
-
-        caseData.setTseViewApplicationOpenOrClosed(OPEN_STATE);
-        caseData.setTseViewApplicationSelect(TseHelper.populateOpenOrClosedApplications(caseData));
-        caseData.getTseViewApplicationSelect().setValue(DynamicValueType.create("1", ""));
-        TseHelper.setDataForTseApplicationSummaryAndResponses(caseData);
-        String expected = "|Application | |\r\n" +
-                "|--|--|\r\n" +
-                "|Applicant | Claimant|\r\n" +
-                "|Type of application | Withdraw my claim|\r\n" +
-                "|Application date | 13 December 2022|\r\n" +
-                "|What do you want to tell or ask the tribunal? | Text|\r\n" +
-                "|Supporting material | N/A|\r\n" +
-                "|Do you want to copy this correspondence to the other party to satisfy the Rules of Procedure? | N/A |\r\n\r\n";
-        assertThat(caseData.getTseApplicationSummaryAndResponsesMarkup(), is(expected));
-    }
-
-    @Test
-    public void setDataForTseApplicationSummaryAndResponses_withAnApplicationToPostpone_setsApplicationSummary() {
-        caseData.getGenericTseApplicationCollection().get(0).getValue().setType(TSE_APP_POSTPONE_A_HEARING);
-        caseData.setTseViewApplicationOpenOrClosed(OPEN_STATE);
-        caseData.setTseViewApplicationSelect(TseHelper.populateOpenOrClosedApplications(caseData));
-        caseData.getTseViewApplicationSelect().setValue(DynamicValueType.create("1", ""));
-        TseHelper.setDataForTseApplicationSummaryAndResponses(caseData);
-        String expected = "|Application | |\r\n" +
-                "|--|--|\r\n" +
-                "|Applicant | Claimant|\r\n" +
-                "|Type of application | Postpone a hearing|\r\n" +
-                "|Application date | 13 December 2022|\r\n" +
-                "|What do you want to tell or ask the tribunal? | Text|\r\n" +
-                "|Supporting material | N/A|\r\n" +
-                "|Do you want to copy this correspondence to the other party to satisfy the Rules of Procedure? | N/A |\r\n" +
-                "\r\n";
-
-        assertThat(caseData.getTseApplicationSummaryAndResponsesMarkup(), is(expected));
-    }
-
-    @Test
-    public void setDataForTseApplicationSummaryAndResponses_withApplicationWithDocument_setsApplicationSummary() {
-        UploadedDocumentType documentType =
-                UploadedDocumentBuilder.builder().withFilename("image.png").withUuid("1234").build();
-        caseData.getGenericTseApplicationCollection().get(0).getValue().setDocumentUpload(documentType);
-
-        caseData.setTseViewApplicationSelect(TseHelper.populateOpenOrClosedApplications(caseData));
-        caseData.getTseViewApplicationSelect().setValue(DynamicValueType.create("1", ""));
-        TseHelper.setDataForTseApplicationSummaryAndResponses(caseData);
-        String expected = "|Application | |\r\n|--|--|\r\n|Applicant | Claimant|\r\n|"+
-                "Type of application | Withdraw my claim|\r\n|Application date | 13 December 2022|\r\n"+
-                "|What do you want to tell or ask the tribunal? | Text|\r\n"+
-                "|Supporting material | <a href=\"/documents/1234/binary\" target=\"_blank\">image.png</a>|\r\n"+
-                "|Do you want to copy this correspondence to the other party to satisfy the Rules of Procedure? | N/A |\r\n\r\n";
-
-        assertThat(caseData.getTseApplicationSummaryAndResponsesMarkup(), is(expected));
-    }
-
-    private UploadedDocumentType createUploadedDocumentType(String fileName) {
-        return UploadedDocumentBuilder.builder()
-            .withFilename(fileName)
-            .withUuid("1234")
-            .build();
-    }
-
-    @Test
-    public void somethingSomething() {
-        TseRespondType tseRespondType = TseRespondType.builder()
-            .from(CLAIMANT_TITLE)
-            .date("23 December 2022")
-            .response("Response Details")
-            .hasSupportingMaterial(YES)
-            .supportingMaterial(List.of(
-                createDocumentTypeItem("image.png"),
-                createDocumentTypeItem("Form.pdf")))
-            .copyToOtherParty(YES)
-            .build();
-
-        TseRespondTypeItem tseRespondTypeItem = TseRespondTypeItem.builder()
-            .id(UUID.randomUUID().toString())
-            .value(tseRespondType)
-            .build();
-
-        GenericTseApplicationType genericTseApplicationType = TseApplicationBuilder.builder()
-            .withNumber("1")
-            .withType(TSE_APP_AMEND_RESPONSE)
-            .withApplicant(RESPONDENT_TITLE)
-            .withDate("13 December 2022")
-            .withDocumentUpload(createUploadedDocumentType("document.txt"))
-            .withDetails("Details Text")
-            .withCopyToOtherPartyYesOrNo(YES)
-            .withStatus(OPEN_STATE)
-            .withRespondCollection(List.of(tseRespondTypeItem))
-            .build();
-
-        GenericTseApplicationTypeItem genericTseApplicationTypeItem = GenericTseApplicationTypeItem.builder()
-            .id(UUID.randomUUID().toString())
-            .value(genericTseApplicationType)
-            .build();
-
-        caseData.setGenericTseApplicationCollection(
-            List.of(genericTseApplicationTypeItem)
-        );
-
-        caseData.setTseViewApplicationSelect(
-            DynamicFixedListType.of(DynamicValueType.create("1", "1 - Amend response")));
-
-        TseHelper.setDataForTseApplicationSummaryAndResponses(caseData);
-
-        String fileDisplay1 = "<a href=\"/documents/1234/binary\" target=\"_blank\">document.txt</a>";
-        String fileDisplay2 = "<a href=\"/documents/1234/binary\" target=\"_blank\">image.png</a>";
-        String fileDisplay3 = "<a href=\"/documents/1234/binary\" target=\"_blank\">Form.pdf</a>";
-
-        String expected = "|Application | |\r\n"
-            + "|--|--|\r\n"
-            + "|Applicant | Respondent|\r\n"
-            + "|Type of application | Amend response|\r\n"
-            + "|Application date | 13 December 2022|\r\n"
-            + "|What do you want to tell or ask the tribunal? | Details Text|\r\n"
-            + "|Supporting material | " + fileDisplay1 + "|\r\n"
-            + "|Do you want to copy this correspondence to the other party to satisfy the Rules of Procedure? "
-            + "| Yes |\r\n\r\n"
-            + "|Responses | |\r\n|--|--|\r\n\r\n"
-            + "|Response 1 | |\r\n"
-            + "|--|--|\r\n"
-            + "|Response from | Claimant|\r\n"
-            + "|Response date | 23 December 2022|\r\n"
-            + "|What’s your response to the respondent’s application? | Response Details|\r\n"
-            + "|Supporting material | " + fileDisplay2 + "<br>" + fileDisplay3 + "<br>" + "|\r\n"
-            + "|Do you want to copy this correspondence to the other party to satisfy the Rules of Procedure? "
-            + "| Yes|\r\n"
-            + "\r\n";
-        
-        assertThat(caseData.getTseApplicationSummaryAndResponsesMarkup(), is(expected));
-
     }
 }
